Move checkpoint methods into a separate container_checkpoint file.
Docker-DCO-1.1-Signed-off-by: Ross Boucher <rboucher@gmail.com> (github: boucher)
